move via-in-pad eligibility into solver
